style: :art: NetlinkErrorを共通化
